+++ lang/ko/messages.php
@@ -11,7 +11,7 @@ return [
         'random' => '랜덤 문서',
         'other' => '기타',
         'empty' => '아직 문서가 없습니다.',
-        'tour_created' => '최근 작성 문서',
+        'tour_recent' => '최근 수정',
         'tour_random' => '랜덤 문서',
     ],
     'doc' => [